Extracted anonymous class from DefinitionContentBuilder to separate class
